+++ packages/puer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+- **Breaking:** Rename `Feature.accept` to `Feature.add` (#59)
 - Use `Iterable` instead of `List` in `Feature` (#28)
 - Add `ReadOnlyFeature` (#27)